+++ retdec/decompiler.py
@@ -27,6 +27,17 @@ class Decompiler(Service):
         :type mode: str
         :param target_language: Target high-level language.
         :type target_language: str
+        :param decomp_var_names: Naming style for variables.
+        :type decomp_var_names: str
+        :param decomp_optimizations: Level of optimizations performed by the
+            decompiler.
+        :type decomp_optimizations: str
+        :param decomp_unreach_funcs: Should all functions be decompiled, even
+            if they are not reachable from the main function?
+        :type decomp_unreach_funcs: bool
+        :param decomp_emit_addresses: Should addresses in comments be emitted
+            in the generated code?
+        :type decomp_emit_addresses: bool
         :param architecture: Architecture. The precise meaning depends on the
             used `mode`.
         :type architecture: str
@@ -80,6 +91,10 @@ class Decompiler(Service):
             'mode': self._get_mode_param(files['input'], kwargs)
         }
         self._add_param_when_given('target_language', params, kwargs)
+        self._add_param_when_given('decomp_var_names', params, kwargs)
+        self._add_param_when_given('decomp_optimizations', params, kwargs)
+        self._add_param_when_given('decomp_unreach_funcs', params, kwargs)
+        self._add_param_when_given('decomp_emit_addresses', params, kwargs)
         self._add_param_when_given('architecture', params, kwargs)
         self._add_param_when_given('file_format', params, kwargs)
         self._add_param_when_given('comp_compiler', params, kwargs)