+++ tests/decompiler_tests.py
@@ -133,6 +133,42 @@ class DecompilerStartDecompilationTests(BaseServiceTests):
             params=AnyParamsWith(target_language='py')
         )
 
+    def test_decomp_var_names_is_set_to_correct_value_when_given(self):
+        self.start_decompilation_with_any_input_file(
+            decomp_var_names='simple'
+        )
+
+        self.assert_post_request_was_sent_with(
+            params=AnyParamsWith(decomp_var_names='simple')
+        )
+
+    def test_decomp_optimizations_is_set_to_correct_value_when_given(self):
+        self.start_decompilation_with_any_input_file(
+            decomp_optimizations='none'
+        )
+
+        self.assert_post_request_was_sent_with(
+            params=AnyParamsWith(decomp_optimizations='none')
+        )
+
+    def test_decomp_unreach_funcs_is_set_to_correct_value_when_given(self):
+        self.start_decompilation_with_any_input_file(
+            decomp_unreach_funcs=True
+        )
+
+        self.assert_post_request_was_sent_with(
+            params=AnyParamsWith(decomp_unreach_funcs=True)
+        )
+
+    def test_decomp_emit_addresses_is_set_to_correct_value_when_given(self):
+        self.start_decompilation_with_any_input_file(
+            decomp_emit_addresses=False
+        )
+
+        self.assert_post_request_was_sent_with(
+            params=AnyParamsWith(decomp_emit_addresses=False)
+        )
+
     def test_architecture_is_set_to_correct_value_when_given(self):
         self.start_decompilation_with_any_input_file(
             architecture='arm'