+++ tests/tools/decompiler_tests.py
@@ -388,6 +388,30 @@ class ParseArgsTests(ToolTestsBase):
         args = parse_args(['decompiler.py', '--compiler-strip', 'file.c'])
         self.assertTrue(args.comp_strip)
 
+    def test_decomp_var_names_is_parsed_correctly_long_form(self):
+        args = parse_args(['decompiler.py', '--var-names', 'simple', 'prog.exe'])
+        self.assertEqual(args.decomp_var_names, 'simple')
+
+    def test_decomp_optimizations_is_parsed_correctly_short_form(self):
+        args = parse_args(['decompiler.py', '-O', 'none', 'prog.exe'])
+        self.assertEqual(args.decomp_optimizations, 'none')
+
+    def test_decomp_optimizations_is_parsed_correctly_long_form(self):
+        args = parse_args(['decompiler.py', '--optimizations', 'none', 'prog.exe'])
+        self.assertEqual(args.decomp_optimizations, 'none')
+
+    def test_decomp_unreach_funcs_is_parsed_correctly_short_form(self):
+        args = parse_args(['decompiler.py', '-K', 'prog.exe'])
+        self.assertTrue(args.decomp_unreach_funcs)
+
+    def test_decomp_unreach_funcs_is_parsed_correctly_long_form(self):
+        args = parse_args(['decompiler.py', '--keep-unreach-funcs', 'prog.exe'])
+        self.assertTrue(args.decomp_unreach_funcs)
+
+    def test_decomp_emit_addresses_is_parsed_correctly_long_form(self):
+        args = parse_args(['decompiler.py', '--no-addresses', 'prog.exe'])
+        self.assertFalse(args.decomp_emit_addresses)
+
     def test_generate_archive_is_set_to_false_when_with_archive_not_given(self):
         args = parse_args(['decompiler.py', 'prog.exe'])
         self.assertFalse(args.generate_archive)
@@ -622,6 +646,42 @@ class MainTests(ToolTestsBase):
 
         self.assert_decompilation_was_started_also_with(
             comp_strip=True
+        )
+
+    def test_sets_decomp_var_names_when_given(self):
+        self.call_main_with_standard_arguments_and(
+            '--var-names', 'simple'
+        )
+
+        self.assert_decompilation_was_started_also_with(
+            decomp_var_names='simple'
+        )
+
+    def test_sets_decomp_optimizations_when_given(self):
+        self.call_main_with_standard_arguments_and(
+            '--optimizations', 'none'
+        )
+
+        self.assert_decompilation_was_started_also_with(
+            decomp_optimizations='none'
+        )
+
+    def test_sets_decomp_unreach_funcs_when_given(self):
+        self.call_main_with_standard_arguments_and(
+            '--keep-unreach-funcs'
+        )
+
+        self.assert_decompilation_was_started_also_with(
+            decomp_unreach_funcs=True
+        )
+
+    def test_sets_decomp_emit_addresses_when_given(self):
+        self.call_main_with_standard_arguments_and(
+            '--no-addresses'
+        )
+
+        self.assert_decompilation_was_started_also_with(
+            decomp_emit_addresses=False
         )
 
     def test_generates_and_saves_archive_when_requested(self):